requirements.txt
# OpenAI Agents SDK and dependencies
openai>=1.50.0
python-dotenv>=1.0.0

# Async support
aiofiles>=24.0.0

# CLI enhancements
prompt-toolkit>=3.0.0
rich>=13.0.0

# Git operations (for validation)
gitpython>=3.1.0

# Data validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
